Statistik Optik angepasst

+++ src/PuzzleStatistics.java
@@ -51,7 +51,7 @@ public class PuzzleStatistics {
     public void printStatistics(int numberOfPuzzles){
         String output = "Heuristic: " + this.heuristic + "\nTotal Number of expanded Nodes: " + this.totalExpandedNodes
                         + "\nTotal Depth: " + this.depth + "\nTotal Computation Time (in ms): " + this.computationTimeMillis + "ms";
-        output += "\nAverage Number of expanded Nodes: " + (this.totalExpandedNodes/numberOfPuzzles) + "\nAverage Depth: " + (this.depth/numberOfPuzzles) + "\nAverage Computation Time (in ms): " + (this.computationTimeMillis/numberOfPuzzles) + "ms";
+        output += "\nAverage Number of expanded Nodes: " + (((double) this.totalExpandedNodes)/((double) numberOfPuzzles)) + "\nAverage Depth: " + (((double) this.depth)/((double) numberOfPuzzles)) + "\nAverage Computation Time (in ms): " + (((double) this.computationTimeMillis)/((double) numberOfPuzzles)) + "ms";
         System.out.println(output);
     }
 }
